requirements.txt
django-money
django-allauth
django-crispy-forms
django-image-cropping
easy-thumbnails
